moved routes for wizard into own routine to keep from cluttering AppWrapper [#155328208]
